refactor: mypage 스타일 구조 개선 및 파일 분리
